Setup post route
