+++ package.json
@@ -18,6 +18,8 @@
   },
   "homepage": "https://github.com/uniqueayo1988/mongo-todo#readme",
   "dependencies": {
+    "body-parser": "^1.19.0",
+    "express": "^4.17.1",
     "mongodb": "^3.5.1",
     "mongoose": "^5.8.9"
   }
