+++ server/models/todo.js
@@ -1,0 +1,21 @@
+const mongoose = require('mongoose')
+
+// Todo Model
+var Todo = mongoose.model('Todo', {
+  text: {
+    type: String,
+    required: true,
+    minlength: 5,
+    trim: true
+  },
+  completed: {
+    type: Boolean,
+    default: false
+  },
+  completedAt: {
+    type: Number,
+    default: null
+  }
+})
+
+module.exports = {Todo}
