+++ server/models/user.js
@@ -1,0 +1,13 @@
+const mongoose = require('mongoose')
+
+// User Model
+var User = mongoose.model('User', {
+  email: {
+    type: String,
+    required: true,
+    minlength: 1,
+    trim: true
+  }
+})
+
+module.exports = {User}
